created ClothingSuggestionTest and modified UI test
moved all of the clothing suggestion testing to its own file because it no longer resides in the UserInterface class
