+++ src/test/UserInterfaceTest.java
@@ -10,6 +10,7 @@ import java.util.Set;
 
 import org.junit.Test;
 
+import project.ClothingSuggestion;
 import project.UserInterface;
 
 /**
@@ -100,155 +101,6 @@ public class UserInterfaceTest {
 		String expectedResult = "could not find";
 		if (!expectedResult.equals(ui.formatTemp(testItem))) {
 			fail("Expected "+ expectedResult + "for test item "+ testItem + "but got " + ui.formatTemp(testItem));
-		}
-	}
-	
-	
-	@Test
-	public void testTempSuggestionHot() {
-		UserInterface ui = new UserInterface();
-		String testTemp = "25";
-		String expectedResult = "It's hot! Try wearing shorts and T-shirt.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTemp))){
-			fail("Expected " + expectedResult + " for test item " + testTemp + " but got " + ui.getTempSuggestion(testTemp));
-		}
-	}
-	@Test
-	public void testTempSuggestionWarm() {
-		UserInterface ui = new UserInterface();
-		String testTempMax = "24";
-		String testTempMin = "20";
-		String expectedResult = "It's warm. Try wearing a T-shirt with shorts or pants.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMax))){
-			fail("Expected " + expectedResult + " for test item " + testTempMax + " but got " + ui.getTempSuggestion(testTempMax));
-		}
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMin))){
-			fail("Expected " + expectedResult + " for test item " + testTempMin + " but got " + ui.getTempSuggestion(testTempMin));
-		}
-	}
-	@Test
-	public void testTempSuggestionAmbient() {
-		UserInterface ui = new UserInterface();
-		String testTempMax = "19";
-		String testTempMin = "15";
-		String expectedResult = "Try wearing pants and a lightweight jacket.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMax))){
-			fail("Expected " + expectedResult + " for test item " + testTempMax + " but got " + ui.getTempSuggestion(testTempMax));
-		}
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMin))){
-			fail("Expected " + expectedResult + " for test item " + testTempMin + " but got " + ui.getTempSuggestion(testTempMin));
-		}
-	}
-	@Test
-	public void testTempSuggestionCool() {
-		UserInterface ui = new UserInterface();
-		String testTempMax = "14";
-		String testTempMin= "10";
-		String expectedResult = "It's cool.  Wear pants and a jacket.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMax))){
-			fail("Expected " + expectedResult + " for test item " + testTempMax + " but got " + ui.getTempSuggestion(testTempMax));
-		}
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMin))){
-			fail("Expected " + expectedResult + " for test item " + testTempMin + " but got " + ui.getTempSuggestion(testTempMin));
-		}
-	}
-	@Test
-	public void testTempSuggestionCold() {
-		UserInterface ui = new UserInterface();
-		String testTempMax = "5";
-		String testTempMin = "0";
-		String expectedResult = "It's cold.  You should wear warm jacket.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMax))){
-			fail("Expected " + expectedResult + " for test item " + testTempMax + " but got " + ui.getTempSuggestion(testTempMax));
-		}
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempMin))){
-			fail("Expected " + expectedResult + " for test item " + testTempMin + " but got " + ui.getTempSuggestion(testTempMin));
-		}
-	}
-	@Test
-	public void testTempSuggestionFreeze() {
-		UserInterface ui = new UserInterface();
-		String testTemp = "-1";
-		String expectedResult = "It's freezing!.  Make sure to wear a winter jacket.";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTemp))){
-			fail("Expected " + expectedResult + " for test item " + testTemp + " but got " + ui.getTempSuggestion(testTemp));
-		}
-	}
-	@Test
-	public void testTempSuggestionEmptyOrNull() {
-		UserInterface ui = new UserInterface();
-		String testTemp = "";
-		String testTempNull = null;
-		String expectedResult = "cannot get clothing suggestion";
-		if(!expectedResult.equals(ui.getTempSuggestion(testTemp))){
-			fail("Expected " + expectedResult + " for test item " + testTemp + " but got " + ui.getTempSuggestion(testTemp));
-		}
-		if(!expectedResult.equals(ui.getTempSuggestion(testTempNull))){
-			fail("Expected " + expectedResult + " for test item " + testTempNull + " but got " + ui.getTempSuggestion(testTempNull));
-		}
-	}
-	
-	@Test
-	public void testStateSuggestionRainyWeather() {
-		UserInterface ui = new UserInterface();
-		String testItemThunder = "t";
-		String testItemHeavyRain = "hr";
-		String testItemLightRain = "lr";
-		String testItemShowers = "s";
-		
-		Set<String> rainRes = new HashSet<String>();
-		rainRes.add("Bring an umbrella!");
-		rainRes.add("Wear a raincoat!");
-		
-		if (!rainRes.contains(ui.getStateSuggestion(testItemThunder))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemThunder + " but got " + ui.getStateSuggestion(testItemThunder));
-		}
-		if (!rainRes.contains(ui.getStateSuggestion(testItemHeavyRain))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemHeavyRain + " but got " + ui.getStateSuggestion(testItemHeavyRain));
-
-		}
-		if (!rainRes.contains(ui.getStateSuggestion(testItemLightRain))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemLightRain + " but got " + ui.getStateSuggestion(testItemLightRain));
-
-		} 
-		if (!rainRes.contains(ui.getStateSuggestion(testItemShowers))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemShowers + " but got " + ui.getStateSuggestion(testItemShowers));
-
-		} 
-	}
-	@Test
-	public void testStateSuggestionWinterWeather() {
-		UserInterface ui = new UserInterface();
-		String testItemSnow = "sn";
-		String testItemSleet = "sl";
-		String testItemHail = "h";
-		
-		Set<String> snowRes = new HashSet<String>();
-		snowRes.add("Wear a hat and gloves!");
-		snowRes.add("Wear winter boots!");
-		
-		if (!snowRes.contains(ui.getStateSuggestion(testItemSnow))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemSnow + " but got " + ui.getStateSuggestion(testItemSnow));
-		}
-		if (!snowRes.contains(ui.getStateSuggestion(testItemSleet))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemSleet + " but got " + ui.getStateSuggestion(testItemSleet));
-
-		}
-		if (!snowRes.contains(ui.getStateSuggestion(testItemHail))) {
-			fail("Expected output in \"snowRes\" for test item " + testItemHail + " but got " + ui.getStateSuggestion(testItemHail));
-
-		} 
-	}
-	@Test
-	public void testStateSuggestionNoSuggestion() {
-		UserInterface ui = new UserInterface();
-		String testItemEmpty = "";
-		String testItemNull = null;
-		if (!ui.getStateSuggestion(testItemEmpty).equals("unable to get suggestion")) {
-			fail("Expected output \"unable to get suggestion\" for test item " + testItemEmpty + " but got " + ui.getStateSuggestion(testItemEmpty));
-		}
-		if (!ui.getStateSuggestion(testItemNull).equals("unable to get suggestion")) {
-			fail("Expected output \"unable to get suggestion\" for test item " + testItemNull + " but got " + ui.getStateSuggestion(testItemNull));
 		}
 	}
